+++ blueprints/ember-select-2/index.js
@@ -2,8 +2,6 @@ module.exports = {
   normalizeEntityName: function() {},
 
   afterInstall: function() {
-    var that = this;
- 
     return this.addBowerPackageToProject('select2');
   }
 };