+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added option to disable the search field for single select dropdowns (#29)
 * Fixed promises in value binding (#30)
 * Change event does not trigger an autorun anymore (#33)
+* Make select2 assertion ember 1.10 proof (#34, thanks, lucas-clemente!)
 
 ### 1.1.0 (November 18, 2014)