+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fixed promises in value binding (#30)
 * Change event does not trigger an autorun anymore (#33)
 * Make select2 assertion ember 1.10 proof (#34, thanks, lucas-clemente!)
+* Added option to customize group headers (#46, thanks, gopeter)
 
 ### 1.1.0 (November 18, 2014)