+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Change event does not trigger an autorun anymore (#33)
 * Make select2 assertion ember 1.10 proof (#34, thanks, lucas-clemente!)
 * Added option to customize group headers (#46, thanks, gopeter)
+* Override default option of select2 libraries for looking up object id (#39, thanks, tylr & kevin-brown)
 
 ### 1.1.0 (November 18, 2014)