+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # ember-select-2 changelog
 
+### 1.2.0 (TBB)
+
+* Added option to disable the search field for single select dropdowns (#29)
+* Fixed promises in value binding (#30)
+* Change event does not trigger an autorun anymore (#33)
+* Make select2 assertion ember 1.10 proof (#34, thanks, lucas-clemente!)
+* Added option to customize group headers (#46, thanks, gopeter)
+* Override default option of select2 libraries for looking up object id (#39, thanks, tylr & kevin-brown)
+* Added `minimumInputLength` option and `didSelect` action (#61, thanks, jniechcial)
+* Added `optionIdPath` option (#58, thanks, ziggythehamster)
+
 ### 1.1.0 (November 18, 2014)
 
 * #18 Allow to specify optionLabelPath instead of relying on text. (thanks, abuiles!)